refactor(api): Move get-api-rate-limit use case to rate-limiting module per PR feedback
