+++ apps/api/src/app/environments/usecases/index.ts
@@ -6,7 +6,6 @@ import { GetEnvironment } from './get-environment';
 import { GetMyEnvironments } from './get-my-environments/get-my-environments.usecase';
 import { UpdateEnvironment } from './update-environment/update-environment.usecase';
 import { GetMxRecord } from '../../inbound-parse/usecases/get-mx-record/get-mx-record.usecase';
-import { GetApiRateLimit } from './get-api-rate-limit';
 
 export const USE_CASES = [
   //
@@ -18,5 +17,4 @@ export const USE_CASES = [
   RegenerateApiKeys,
   GetEnvironment,
   GetMyEnvironments,
-  GetApiRateLimit,
 ];
