+++ apps/api/src/app/rate-limiting/usecases/get-api-rate-limit/get-api-rate-limit.spec.ts
@@ -7,9 +7,10 @@ import { Test } from '@nestjs/testing';
 import { CacheService, MockCacheService } from '@novu/application-generic';
 import { GetApiRateLimit, GetApiRateLimitCommand } from './index';
 import { SharedModule } from '../../../shared/shared.module';
-import { EnvironmentsModule } from '../../environments.module';
+import { EnvironmentsModule } from '../../../environments/environments.module';
 import { OrganizationModule } from '../../../organization/organization.module';
-import { GetDefaultApiRateLimits } from '../../../rate-limiting/usecases/get-default-api-rate-limits';
+import { GetDefaultApiRateLimits } from '../get-default-api-rate-limits';
+import { RateLimitingModule } from '../../rate-limiting.module';
 
 const mockDefaultApiRateLimits = {
   [ApiServiceLevelTypeEnum.FREE]: {
@@ -37,7 +38,7 @@ describe('GetApiRateLimit', async () => {
 
   beforeEach(async () => {
     const moduleRef = await Test.createTestingModule({
-      imports: [SharedModule, EnvironmentsModule, OrganizationModule],
+      imports: [SharedModule, RateLimitingModule],
       providers: [],
     })
       .overrideProvider(CacheService)
